+++ Commands/Base/SPOnlineConnection.cs
@@ -153,12 +153,12 @@ namespace SharePointPnP.PowerShell.Commands.Base
 
         public void RestoreCachedContext(string url)
         {
-            Context = ContextCache.FirstOrDefault(c => HttpUtility.UrlEncode(c.Url) == HttpUtility.UrlEncode(url));
+            Context = ContextCache.FirstOrDefault(c => new Uri(c.Url).AbsoluteUri == new Uri(url).AbsoluteUri);
         }
 
         internal void CacheContext()
         {
-            var c = ContextCache.FirstOrDefault(cc => HttpUtility.UrlEncode(cc.Url) == HttpUtility.UrlEncode(Context.Url));
+            var c = ContextCache.FirstOrDefault(cc => new Uri(cc.Url).AbsoluteUri == new Uri(Context.Url).AbsoluteUri);
             if (c == null)
             {
                 ContextCache.Add(Context);
@@ -167,7 +167,7 @@ namespace SharePointPnP.PowerShell.Commands.Base
 
         public ClientContext CloneContext(string url)
         {
-            var context = ContextCache.FirstOrDefault(c => HttpUtility.UrlEncode(c.Url) == HttpUtility.UrlEncode(url));
+            var context = ContextCache.FirstOrDefault(c => new Uri(c.Url).AbsoluteUri == new Uri(url).AbsoluteUri);
             if (context == null)
             {
                 context = Context.Clone(url);
@@ -176,8 +176,8 @@ namespace SharePointPnP.PowerShell.Commands.Base
                     context.ExecuteQueryRetry();
                 }
                 catch (Exception ex)
-
                 {
+#if !ONPREMISES && !NETSTANDARD2_0
                     if (ex is WebException || ex is NotSupportedException)
                     {
                         // legacy auth?
@@ -187,8 +187,11 @@ namespace SharePointPnP.PowerShell.Commands.Base
                     }
                     else
                     {
+#endif
                         throw;
+#if !ONPREMISES && !NETSTANDARD2_0
                     }
+#endif
                 }
                 ContextCache.Add(context);
             }