+++ Commands/Branding/DisableResponsiveUI.cs
@@ -5,7 +5,7 @@ using SharePointPnP.PowerShell.CmdletHelpAttributes;
 namespace SharePointPnP.PowerShell.Commands.Branding
 {
     [Cmdlet(VerbsLifecycle.Disable, "PnPResponsiveUI")]
-    [CmdletHelp("Deactive the PnP Response UI add-on",
+    [CmdletHelp("Deactivate the PnP Response UI add-on",
         "Disables the PnP Responsive UI implementation on a classic SharePoint Site", Category = CmdletHelpCategory.Branding)]
     [CmdletExample(Code = "PS:> Disable-PnPResponsiveUI",
         Remarks = @"If enabled previously, this will remove the PnP Responsive UI from a site.",