+++ Commands/Files/GetFile.cs
@@ -105,11 +105,11 @@ namespace SharePointPnP.PowerShell.Commands.Files
             {
                 case URLTOPATH:
 
-                    SaveFileToLocal(SelectedWeb, serverRelativeUrl, Path, Filename, (filetosave) =>
+                    SaveFileToLocal(SelectedWeb, serverRelativeUrl, Path, Filename, (fileToSave) =>
                     {
                         if (!Force)
                         {
-                            WriteWarning($"File '{filetosave}' exists already. use the -Force parameter to overwrite the file.");
+                            WriteWarning($"File '{fileToSave}' exists already. use the -Force parameter to overwrite the file.");
                         }
                         return Force;
                     });
@@ -158,7 +158,12 @@ namespace SharePointPnP.PowerShell.Commands.Files
 
         private void SaveFileToLocal(Web web, string serverRelativeUrl, string localPath, string localFileName = null, Func<string, bool> fileExistsCallBack = null)
         {
+
+#if SP2013 || SP2016
+            var file = web.GetFileByServerRelativeUrl(serverRelativeUrl);
+#else
             var file = web.GetFileByServerRelativePath(ResourcePath.FromDecodedUrl(serverRelativeUrl));
+#endif
 
             var clientContext = web.Context as ClientContext;
             clientContext.Load(file);