+++ CHANGELOG.md
@@ -5,19 +5,39 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [3.8.1904.0] - unreleased 
+
+### Added
+
+### Changed
+
+### Contributors
+
 ## [3.7.1903.0]
 
 ### Added
+- Added support for client side pages on SP2019
+- Added support for ALM cmdlets (Add-PnPApp, Get-PnPApp etc.) on SP2019
+- Added Add-PnPAlert cmdlet to create alerts (SPO and SP2019 only)
+- Added Get-PnPAlert to list alerts (SPO and SP2019 only)
+- Added Remove-PnPAlert to remove alerts (SPO and SP2019 only)
 - Added support to Connect-PnPOnline authenticate to SharePoint Online when Legacy Authentication has been turned off (Set-PnPTenant -LegacyAuthProtocolsEnabled:$false / Set-SPOTenant -LegacyAuthProtocolsEnabled:$false)
+- Support for cross site page transformation (create modern pages in other site then the one hosting the classic pages) via the TargetWebUrl parameter
+- Support for page transformation mapping parameters (UseCommunityScriptEditor and SummaryLinksToHtml). The first one will use the community script editor as a possible modern target web part, the second one will transform the summarylinks web part to html text instead of the default QuickLinks web part
 
 ### Changed
 - Many typo fixes in code
 - Apply-PnPTenantTemplate will now list the sites created after applying a tenant template.
 - Fixed an issue with Connect-PnPOnline throwing an exception when authenticating using the -SPOManagementShell parameter.
+- Fixed connection issue with URL's containing spaces - #1250
 
 ### Contributors
 
+- Heinrich Ulbricht [heinrich-ulbricht] 
 - Nick Schonning [nschonni]
+- Koen Zomers [KoenZomers]
+- Marvin Dickhaus [Weishaupt]
+- Lars Fernhomberg [lafe]
 
 ## [3.6.1902.2]